Migrate passage note tags store to Pinia
